requirements.txt
click
requests
python-dotenv
rich
